Fix or silence (msvc) compiler warnings about narrowing conversions, unused formal parameters, constant conditional expressions.
(cherry picked from commit 8eae9553595dcfd9bf7802e946ff49ea07405b35)

+++ test/algorithms/overlay/get_turn_info.cpp
@@ -184,7 +184,7 @@ void test_both(std::string const& caseid,
     //return;
 
     std::string reverse;
-    for (int i = expected.size() - 1; i >= 0; i--)
+    for (int i = static_cast<int>(expected.size() - 1); i >= 0; i--)
     {
         reverse += expected[i];
     }
