+++ test/multi/algorithms/multi_num_points.cpp
@@ -25,7 +25,7 @@
 template <typename Geometry>
 void check_geometry(Geometry const& geometry, std::string const& wkt, int expected)
 {
-    int detected = bg::num_points(geometry);
+    size_t detected = bg::num_points(geometry);
     BOOST_CHECK_MESSAGE(detected == expected,
         "num_points: " << wkt
         << " -> Expected: " << expected
